+++ docs/source/history.rst
@@ -2,6 +2,7 @@ Release history
 ===============
 
 Version 0.4
+  - ``sample_format`` is a new encoder configuration option.
   - The encoders sample format is 's16le' except for the 'audio/l16' encoder.
   - The encoder command line is now updated with ``pa-dlna.conf`` user
     configuration.